% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataP.R
\docType{data}
\name{dataP}
\alias{dataP}
\title{Simulated data from zero-inflated exponential regression model}
\format{
A data frame which contains x1, x2 and y.
\describe{
  \item{y}{the response variable}
  \item{x1}{Binary covariate}
  \item{x2}{Continuous covariate}
}
}
\usage{
dataP
}
\description{
Simulated data was generated with x1 following a Bernoulli distribution with a success probability of 0.4, x2 following a standard normal distribution, and y following a zero-inflated exponential regression model.
}
\seealso{
\code{\link{UHM}},\code{\link{ZIHR}}
}
\keyword{datasets}
