% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHM.R
\docType{package}
\name{UHM}
\alias{UHM-package}
\alias{UHM}
\title{UHM Package}
\description{
Run a Gibbs sampler for hurdle models. The package includes the hurdle generalized linear model under Gaussian, exponential, Gamma, Weibull, inverse Gaussian, Poisson, negative binomial, logarithmic, logistic, and binomial distributional assumptions. The package also considers hurdle generalized Poisson models and hurdle Beta regression models. For model comparison, Deviance Information Criterion (DIC) and Log Pseudo Marginal Likelihood (LPML) are presented.
}
\references{
\enumerate{
\item
Ganjali, M., Baghfalaki, T. & Balakrishnan, N. (2024). A Unified Bayesian approach for Modeling Zero-Inflated count and continuous outcomes.

}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tbaghfalaki/UHM}
}

}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}, Mojtaba Ganjali \email{m-ganjali@sbu.ac.ir}, Narayanaswamy Balakrishnan \email{bala@mcmaster.ca}
}
\keyword{UHM}
